Modified code so that many of the parameters used in driver_fungalGrowth_singleNutrient.py can now be input through the parameter.ini file. The user then runs the command line: python driver_fungalGrowth_singleNutrient.py -i paramater_file_name
New parameters are specified in the new parameter file. However, the code should be backwards compatible to also allow the parameter file to be specified in code.
Also, if a parameter is not in the parameter file, there is a default set. The default parameter can be changed in the code.